pyproject.toml
[tool.poetry]
name = "python-rest-template"
version = "0.1.0"
description = ""
authors = ["Jakob J. Kolb <jakob.j.kolb@gmail.com>"]
readme = "README.md"
packages = [{include = "python_rest_template"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.89.1"}
yoyo-migrations = "^8.2.0"
ramda = "^0.7.6"
psycopg2-binary = "^2.9.5"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
pytest-xdist = "^3.1.0"
black = "^22.12.0"
uvicorn = "^0.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
